+++ packages/ard/index.js
@@ -1,0 +1,7 @@
+// import packages
+const createHashedId = require('../../utils/ard/createHashedId')
+
+// export packages
+module.exports = {
+	createHashedId,
+}