+++ packages/storage-wrapper/createUri.js
@@ -9,10 +9,10 @@
 module.exports = {
 	s3: (bucket, path) => {
 		// return data
-		return ['s3:/', bucket, path].join('/');
+		return ['s3:/', bucket, path].join('/')
 	},
 	gs: (bucket, path) => {
 		// return data
-		return ['gs:/', bucket, path].join('/');
+		return ['gs:/', bucket, path].join('/')
 	},
 }