+++ utils/strings/isIncluded.js
@@ -1,0 +1,2 @@
+// check if a variable (haystack) includes a needle
+module.exports = (haystack, needle) => haystack.indexOf(needle) !== -1