+++ utils/strings/toHex.js
@@ -1,0 +1,2 @@
+// convert input to hex value
+module.exports = (value) => Buffer.from(value).toString('hex')