+++ CHANGELOG.md
@@ -1,15 +1,22 @@
-# Storage Wrapper for NodeJS
+# SWR Audio Lab / Node Utils Changelog
 
-by [**SWR audio lab**](https://lab.swr.de/)  
+by [**SWR Audio Lab**](https://lab.swr.de/)  
 
 ## Changelog
+
+- 2022-01-26 - v1.0.0-beta
+  - chore!: folder migration
+  - refactor: swap node-fetch for undici
+  - feat: add ard publisher script
+  - feat: add ard category parser
+  - feat: add string helpers
 
 - 2021-08-11 - v0.2.3
   - updated dependencies
   - updated github actions
   - add optional 'resumable' flag
 
-- 2021-06-22 - v0.2.2
+- 2021-06-21 - v0.2.2
   - updated dependencies
   - updated github actions
   - changed dependabot interval